shim ml2 driver_api with neutron-lib's api
neutron-lib contains the ML2 driver_api module and while we'd like to
switch over to lib's version, we have a significant number of consumers
[1]. Therefore this patch proposes to shim to the remaining public
driver_api like we did in I99e3de6d933a1bb341394f85415fb07306a82a01.

A subsequent patch will move neutron's imports over to use lib's version
of the module; this patch is intended to be "self testing" in that it
assures the shim works in neutron (for neutron's gate anyway).

[1] http://codesearch.openstack.org/?q=from%20neutron.plugins.ml2%20import%20driver_api

Change-Id: I86a10091b55d1123e8d16f16155e0312bb10e54c