functional: disable netlink tests
Current gate runs with kernel 4.4.0-89. Creating conntrack entries fails
with this version and leaves conntrack in stuck state leading to
functional tests to hang.

This patch proposes to skip those tests that create conntrack entries so
we can enable back the functional job to voting to avoid new regressions
at this suite.

Change-Id: I9f883cea1d101efe6a7214c3a9b679fc32427365
Related-bug: #1709032