Stop logging versions on every agent update
This logging statement was being called on every agent update
even if no versions changed, resulting in a bunch of useless
messages.

We already log when the actual versions change in the _set_version
call here: [1]

To cover the cases where versions are removed, this adds two logging
statements so we can be sure we see when the tracked versions do
change but not be bothered by version logs in any other condition.

1.
http://git.openstack.org/cgit/openstack/neutron/tree/neutron/api/rpc/callbacks/version_manager.py?h=d225b86738a9db56bc643dfe637b1f21393ad27e#n113

Partial-Bug: #1707307
Change-Id: I4761d21257dabfbd661bba1b24fe09d208d16ae6