Return neutron-ns-metadata-proxy threads to 1000
A wsgi threads change [1] also changed the number of threads the
neutron-ns-metadata-proxy process has down to 100.  At some sites this
low setting is leading to failures.
This patch increases only the neutron-ns-metadata-proxy default thread count
back up to 1000 to resolve these failures.

This is not a cherry pick from Ocata as this code was removed in Ocata
in favor of using HAProxy with a setting of 1024 [2].

[1] https://review.openstack.org/278007/
[2] https://review.openstack.org/431691

Change-Id: I59cb05e165c6daa4ce0a1b636450f66fb699e88e
Closes-Bug: #1711215