Allow to disable DNS server announcement per subnet
Currently there is no way to have DHCP agents not announce DNS servers
for a subnet. The current behaviour when the dns_nameservers option is
set to '0.0.0.0' is that each agent will only announce itself instead of
announcing the list of all dhcp agents for that subnet, which seems not
too useful. So we redefine the meaning of this option to instruct the
DHCP agent to not announce any DNS server in that case.

Actually, going back to square one, it would be more natural to swap the
meaning of "option unset" and "option 0.0.0.0", but that would change
the default behaviour for all existing installation and thus does not
seem feasible.

Change-Id: I32d943360162c483ac1364100a21ab56b13517fb
Closes-Bug: 1311040