DVR: _get_floatingips_bound_to_host throws KeyError
_get_floatingips_bound_to_host function was introduced
recently in dvr_local_router to retrieve the external
interface name for centralizing the floatingip.

This function was throwing a 'KeyError' on fip['host'] and
not required for centralized floatingips anymore.

The get_external_device_interface_name in dvr_local_router
will try to get the 'fg' interface that is required for
the bound floating-ips to clear up some of the rules.
In the case of the centralized unbound floating-ips, the
'qg' external interface is retreived from
get_snat_external_device_interface_name that is defined
in 'dvr_edge_router' and based on the namespace.

So _get_floatingips_bound_to_host can be removed from
get_external_device_inteface_name.

Closes-Bug: 1712412

Change-Id: I94c0a071df32f572745a2c29942956c3da9f309b
(cherry picked from commit 47fbc6157ac4125fa068e64b573433a02c0ce0fc)