Pecan: add plugin pagination/sorting validation
This adds the validation to ensure that the plugin supports
native sorting when native pagination is used.

This patch doesn't add a unit test for this because it will
be covered in the switch to pecan for the existing unit tests
in I76dc23fb7b96d82b0da50285bd0aac76142e81e5 (which is how this
bug was discovered).

Closes-Bug: #1714381
Change-Id: I6443832357c91fe791853a374cdec11dd1f968ea
(cherry picked from commit 31dc80a0c22ae1cbacad1694f799cbdc4228cfd4)