DVR: Fix agent to process only floatingips that have a host match
The agent is not currently checking for the host bound
before configuring the floatingip. That leads to
floatingips being configured on multiple hosts.

This is a partial fix on the agent side to prevent
configuring a floatingip ip that is not bound to
this host.

Related-Bug: #1712412
Related-Bug: #1713927

Change-Id: I1bc8c42425f97234f56412a2f109a996d9f896de
(cherry picked from commit afd1995d91353874d04ac2d12be0162cea27c1d1)