DHCP provisioning block only on port addr update
Don't add a DHCP provisioning block during port updates
unless the IP address or MAC address has changed. There is
no point in waiting on the DHCP agent after a port update
unless the addresses have changed. We already wait on the
DHCP agent after the port has been initially created so
we will be guaranteed that a reservation exists for the
correct MAC/IP in the DHCP agent.

Closes-Bug: #1714655
Change-Id: Idef5994e274ea99260367edb8331fd10b98a1432