Treat lack of segment info in port object as unbound
A push notifications change added segment information
to the get_device_details() RPC call, but sometimes the
segment information is not present, resulting in an
AttributeError.  Just treat the lack of segment info
as if the port was unbound, since the port is probably
in the process of being removed.

Change-Id: I631c6e1f02fa07eed330c99a96aa66d747784f37
Closes-bug: #1714068