fix missing l2pop config option docs
The l2pop ML2 mechanism driver's configuration options are no longer
being generated during the docs build. This patch adds the l2pop
config options back into the ml2 sample config generation and also
fixes a link to them in the admin docs.

This fix is a candidate for backport to pike.

Change-Id: Ia26b4d1995690e7291b6476dc683271e12de09ab
Closes-Bug: #1714528