Don't trigger DVR port update if status the same
The DVR path was missing logic to avoid triggering DB updates
even if the status was already correct. This was leading to a
constant back-and-forth updating between the agent and the server
effectively not changing anything but the updated_at timestamp
and the revision_number.

Closes-Bug: #1716522
Change-Id: I88c678ac8a562db91c71edbc71dc43d1ada6c452
(cherry picked from commit 1edd0c792499dbd6c59257a10c901360941dbcf6)