Stop using is_agent_down in agents_db
The usage of is_agent_down in neutron.db.agents_db.AgentDbMixin
is deprecated for removal in version ocata, and usage of the
is_agent_down neutron.agent.common.utils is recommended in future
release. And there is no usage of is_agent_down via
neutron.db.agents_db.AgentDbMixin in other projects [1]. It's safe
to remove it now.

[1]
http://codesearch.openstack.org/?q=is_agent_down&i=nope&files=&repos=

Change-Id: I7fbd2aa7294f284de66b131b59be9150e0538251