make net_helpers functions work on OpenSUSE
/sbin may not be in the regular user's PATH or tools like sysctl/ss
may require root privileges to execute correctly on OpenSUSE, and this
makes net_helpers functions fail with OSError. There is no harm in
running ss or sysctl as root user for these functions and that allows
fullstack/functional tests to operate correctly on OpenSUSE.

The change requires a testcase to inherit from BaseSudoTestCase due
to the new run_as_root=True flag.

Change-Id: Ia4f2af1d44faacf5f7ab5471b4f18ecb27f06549