Replace default subnetpool API tests with UT
As existence of default subnetpool is required for auto allocated tests
and other tests in Nova, it is created by devstack. Default subnetpool
API tests perform create operations and tests are skipped in case
default subnetpool already exsits in the system.

This patch replaces coverage of skipped API tests, that are never run on
the gate, with unittests using extension testing.

Change-Id: I04986aaf06525db45290b0b838cc491ee99b751f