Stop arping when IP address gets deleted
A recent change decreased the severity of the arping log
message when the interface was found to have been deleted,
but it can also be the case that the interface exists, but
the IP address was deleted concurrently.  Downgrade log
message to info since there is nothing to be done about it.

Change-Id: Ie63d306802d07f7376e87daf5c88672f7d58093d
Related-bug: #1696893