Allow to configure DHCP T1 and T2 timers in dnsmasq
This patch introduces two new options for dnsmasq in neutron-dhcp-agent:

* dhcp_renewal_time (T1): This option specifies the time interval from
  address assignment until the client transitions to the RENEWING
  state.
* dhcp_rebinding_time (T2): This option specifies the time interval
  from address assignment until the client transitions to the REBINDING
  state.

By allowing to set these timers we can configure both the renewal
and rebinding times (options 58 and 59 as per RFC2132) and, for example
allow to change some parameters (like MTU) on instances without having
to wait or the lease time. The advantage of changing T1 over the lease
time is that if the DHCP server becomes unreachable within the lease
time, instances won't drop their IP addresses and won't cause a
dataplane disruption.

Change-Id: I29d417d459e92f36c1077962b92fa4c43dfaa97d
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>