Change QoS configuration manual
As [1] shows, the controller node hosts the Neutron server but
also agents like L3 and DHCP which require also OVS or LinuxBridge
agent to be running on it.
To enable QoS is required to enable the 'service_plugins' and
the 'extension_drivers', along with the agent section in the plugin
config if the agent is running on this host.

In the network node and the compute node only the agent
'extensions' configuration is needed to enable QoS
on the agent.

[1] https://docs.openstack.org/security-guide/networking/architecture.html

Closes-Bug: #1720077

Change-Id: I14128aabe0a9209c31a1bd4c76eed1182364ccdf
Co-Authored-By: Slawek Kaplonski <slawek@kaplonski.pl>