[Qos] ingress bandwidth limit by ovs is not accurate
According to Open vSwitch FAQ [1], max-rate value which is set for
all queues in qos is set as link speed in case when it is not specified.
This can lead to inaccurate ingress bandwidth limits in case when QoS is
applied e.g. directly on tapXXX port (which is "tun" interface type) and
configured bandwidth limit is higher than interface's link_speed.
This patch set max-rate parameter in qos's other_config table to not use
default value determined by Open vSwitch and to make rate limits more
accurate.

The modification is covered by functional tests in
https://github.com/openstack/neutron/blob/master/neutron/tests/
functional/agent/test_ovs_lib.py.

[1] http://docs.openvswitch.org/en/latest/faq/qos/

Change-Id: Id937216a724fbf776298efd11f74ac71056cfe06
Closes-Bug: #1730605