Treat device with no MAC address as not ready
There was a recent traceback in the dhcp-agent where
it's tap device did not have a MAC address.  Instead
of failing in init_l3(), have the dhcp-agent fail in
setup() where it verifies the device it is about to
use is ready for use, that way it will release the
port and possibly retry.

Closes-bug: 1742355

Change-Id: Id37396b591946498cdc111eaf4f0c5792e67c53a