Don't create new QoS policy during update operation
The QoS plugin erroneously creates a new policy when updating
an existing one.
This causes the loss of previously set attributes that are
not passed in the update call.

This patch fixes this issue by getting existing qos policy object
instead of creating a new one for update.

This issue is already fixed in Pike branch with patch:
I57011bd4d40479855203061554090d19bb668960

Change-Id: I6f5a925084bb70b56e66d673eac2febdb38e0b34
Related-Bug: #1734136